moved functions for handling structure files to ase_tools + cleanup
